+++ website/config.rb
@@ -1,11 +1,11 @@
 set :base_url, "https://www.vagrantup.com/"
 
 set :vmware_utility_name, "vagrant-vmware-utility"
-set :vmware_utility_version, "1.0.2"
+set :vmware_utility_version, "1.0.4"
 
 activate :hashicorp do |h|
   h.name         = "vagrant"
-  h.version      = "2.1.2"
+  h.version      = "2.1.4"
   h.github_slug  = "hashicorp/vagrant"
   h.website_root = "website"
 end