+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 FEATURES:
 
+IMPROVEMENTS:
+
+BUG FIXES:
+
+## 2.1.4 (August 30, 2018)
+
+BUG FIXES:
+
+- core: Fix local plugin installation prompt answer parsing [GH-10154]
+- core: Reset internal environment after plugin loading [GH-10155]
+- host/windows: Fix SMB list parsing when extra fields are included [GH-10156]
+
+## 2.1.3 (August 29, 2018)
+
+FEATURES:
+
 - core: Support for project specific plugins [GH-10037]
 
 IMPROVEMENTS: